new comman line arguments to take shots. new scripts to create shots. lib folder for all scripts.
