+++ libs/index.js
@@ -1,0 +1,9 @@
+import { recordGif as takeGif } from "./gifShot.js";
+import { takeSS as takeScreenshot } from "./takeScreenshot.js";
+import { takeSS as mockupScreenshot } from "./screenshot-with-mockup.js";
+
+export {
+    takeGif,
+    takeScreenshot,
+    mockupScreenshot
+}
