+++ libs/screenshot-with-mockup-sharp.js
